feat(cilium): Enable bpf.hostLegacyRouting for Egress Gateway
This option is essential for Cilium Egress Gateway to function correctly
on specific Kubernetes distributions, such as Talos Linux. It ensures
the host's legacy routing stack is properly integrated with Cilium's BPF
mechanisms when the egress gateway feature is enabled.

This aligns with recommendations found in the Cilium documentation:
https://github.com/cilium/cilium/blob/e2aa6ca523e4f980869a6bfe2ba05d50c9942d98/Documentation/installation/k8s-install-talos-linux.rst

The value for this setting is dynamically configured based on
var.cilium_enable_egress_gateway.

Signed-off-by: Enrique Hernández Bello <enrique.hernandez@maarlab.com>